Continued programming + Simulateur + BarriereEBPP [end TP 1]

+++ 3if/pc/tp-pc_1/src/Mere.cpp
@@ -1,24 +1,58 @@
 #include <stdlib.h>
 #include <unistd.h>
 #include <signal.h>
+#include <sys/types.h>
+#include <sys/wait.h>
+#include <sys/types.h>
+#include <sys/ipc.h>
+#include <sys/msg.h>
+
 #include <Outils.h>
 #include <Heure.h>
 
-int main (int argc, char* argv[])
+#include "Simulateur.h"
+#include "BarriereEBPP.h"
+
+int main ()
 {
+    pid_t pidHeure, pidSimulateur, pidBarriereEBPP;
+    
     //enum TypeTerminal terminal = XTERM;
     enum TypeTerminal terminal = VT220;
 
     /* --- Initialisation --- */
     InitialiserApplication(terminal);
-    pid_t pidHeure = ActiverHeure();
+    pidHeure = ActiverHeure();
 
-    /* --- Moteur --- */
-    sleep(3);
+    key_t cleBarriereEBPP = ftok("/bin/tp-pc_1", 's');
+    int identifiantBarriereEBPP = msgget(cleBarriereEBPP, IPC_CREAT);
     
-    /* --- Destrcution --- */
-    kill(pidHeure, SIGUSR2);
-    TerminerApplication(true);
+    if((pidSimulateur=fork()) == 0)
+    {
+        // Fils - Simulateur
+        Simulateur();
+    }
+    else
+    {
+        // Pere
+        if((pidBarriereEBPP=fork()) == 0)
+        {
+            // Fils - BarriereEBPP            
+            BarriereEBPP();
+        }
+        else
+        {
+            // Pere
+            /* --- Moteur --- */
+            waitpid(pidSimulateur, NULL, 0);
+
+            /* --- Destruction --- */
+            msgctl(identifiantBarriereEBPP, IPC_RMID, NULL);
+            kill(pidHeure, SIGUSR2);
+            TerminerApplication(true);
+            exit(0);
+        }
+    }
 
     return 0;
 }
